Move querySelectors to domUpdates file, remove imports / exports
