+++ src/scripts.js
@@ -20,26 +20,6 @@ import {
   promises
 } from './apiCalls'
 
-// Query Selectors:
-// users
-var userInfoContainer = document.querySelector('.user-info');
-var welcomeHeading = document.querySelector('.welcome-heading');
-var friendList = document.querySelector('.friend-list')
-var gateInfo = document.querySelector('.gate-info')
-// hydration
-var dailyHydrationStats = document.querySelector('.daily-hydration-stats')
-var weeklyHydrationStats = document.querySelector('.weekly-hydration-stats')
-// activity
-var usersStepGoal = document.querySelector('.all-users-step-goal')
-var dailyStepCount = document.querySelector('.daily-steps')
-var dailyActiveMinutes = document.querySelector('.daily-active-minutes')
-var weeklyStepCountGoal = document.querySelector('.weekly-step-count-goal')
-var milesWalkedByDay = document.querySelector('.todays-miles-walked')
-// sleep
-var weeklySleepStats = document.querySelector('.weekly-sleep-stats')
-var sleepStatsByDay = document.querySelector('.today-sleep-stats')
-var avgAllTimeSleepStats = document.querySelector('.all-time-sleep-stats-container')
-
 // Event listener:
 window.addEventListener('load', () => {
   Promise.all(promises)
@@ -63,21 +43,3 @@ window.addEventListener('load', () => {
   })
   .catch(error => console.log('ERROR', error))
 })
-
-export {
-  userInfoContainer,
-  welcomeHeading,
-  friendList,
-  dailyHydrationStats,
-  weeklyHydrationStats,
-  sleepStatsByDay,
-  avgAllTimeSleepStats,
-  dailyStepCount,
-  usersStepGoal,
-  weeklyStepCountGoal,
-  milesWalkedByDay,
-  weeklySleepStats,
-  dailyActiveMinutes,
-  gateInfo
-}
-
